requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
jinja2
openai
python-dotenv
pytest
